composer/blog/gcp-tech-blog/data-orchestration-with-composer/requirements.txt
apache-airflow[gcp]==1.10.14